+++ packages/screenshoter/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+- handle firefox buggy versions
 
 ## 3.0.3 - 2021/1/27